+++ docs/conf.py
@@ -1,0 +1,141 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# For the full list of built-in configuration values, see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+import os
+import sys
+
+# -- Path setup --------------------------------------------------------------
+# Add the project root to sys.path to enable autodoc
+sys.path.insert(0, os.path.abspath(".."))
+
+# -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+project = "ForzaEmbed"
+copyright = "2026, Béranger Thomas"
+author = "Béranger Thomas"
+release = "1.1.0"
+version = "1.1.0"
+
+# -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+extensions = [
+    "sphinx.ext.autodoc",  # Core extension for automatic documentation
+    "sphinx.ext.napoleon",  # Support for Google and NumPy style docstrings
+    "sphinx.ext.viewcode",  # Add links to highlighted source code
+    "sphinx.ext.intersphinx",  # Link to other project's documentation
+    "sphinx.ext.autosummary",  # Generate autodoc summaries
+    "sphinx.ext.coverage",  # Check documentation coverage
+    "sphinx.ext.mathjax",  # Render math via JavaScript
+    "sphinx_rtd_theme",  # Read the Docs theme
+]
+
+# Napoleon settings for Google-style docstrings
+napoleon_google_docstring = True
+napoleon_numpy_docstring = False
+napoleon_include_init_with_doc = True
+napoleon_include_private_with_doc = False
+napoleon_include_special_with_doc = True
+napoleon_use_admonition_for_examples = True
+napoleon_use_admonition_for_notes = True
+napoleon_use_admonition_for_references = False
+napoleon_use_ivar = False
+napoleon_use_param = True
+napoleon_use_rtype = True
+napoleon_preprocess_types = False
+napoleon_type_aliases = None
+napoleon_attr_annotations = True
+
+# Autosummary settings
+autosummary_generate = True
+autosummary_imported_members = False
+
+# Autodoc settings
+autodoc_default_options = {
+    "members": True,
+    "member-order": "bysource",
+    "special-members": "__init__",
+    "undoc-members": True,
+    "exclude-members": "__weakref__",
+}
+autodoc_typehints = "description"
+autodoc_typehints_description_target = "documented"
+
+# Intersphinx mapping
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "numpy": ("https://numpy.org/doc/stable/", None),
+    "sklearn": ("https://scikit-learn.org/stable/", None),
+}
+
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "sav"]
+
+# -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
+
+# Theme options
+html_theme_options = {
+    "prev_next_buttons_location": "bottom",
+    "style_external_links": False,
+    "vcs_pageview_mode": "",
+    "collapse_navigation": False,
+    "sticky_navigation": True,
+    "navigation_depth": 4,
+    "includehidden": True,
+    "titles_only": False,
+}
+
+# Additional HTML options
+html_title = f"{project} {version}"
+html_short_title = project
+html_show_sourcelink = True
+html_show_sphinx = True
+html_show_copyright = True
+
+# Output file base name for HTML help builder
+htmlhelp_basename = "ForzaEmbeddoc"
+
+# -- Options for LaTeX output ------------------------------------------------
+latex_elements = {
+    "papersize": "a4paper",
+    "pointsize": "10pt",
+}
+
+latex_documents = [
+    (
+        "index",
+        "ForzaEmbed.tex",
+        "ForzaEmbed Documentation",
+        "Béranger Thomas",
+        "manual",
+    ),
+]
+
+# -- Options for manual page output ------------------------------------------
+man_pages = [(
+    "index",
+    "forzaembed",
+    "ForzaEmbed Documentation",
+    [author],
+    1,
+)]
+
+# -- Options for Texinfo output ----------------------------------------------
+texinfo_documents = [
+    (
+        "index",
+        "ForzaEmbed",
+        "ForzaEmbed Documentation",
+        author,
+        "ForzaEmbed",
+        "A Python framework for text embedding model evaluation and comparison",
+        "Miscellaneous",
+    ),
+]